Add files via upload
Introduction:
The dataset at hand provides a comprehensive compilation of
information concerning a diverse array of laptops, encompassing a
spectrum of specifications and corresponding prices. Crafted with
the aim of facilitating sales price prediction, this dataset offers
invaluable insights into the intricate details of laptops, ranging from
processor specifications to display attributes. With its rich
repository of data, this dataset emerges as a pivotal resource for
discerning trends within the laptop market and forecasting prices
with precision.